add Pages (Home,Profile)
